+++ RELEASES.md
@@ -8,6 +8,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template-compiler: util for workflows, CompilerBeanHistory and LocalEnactor
 * prov-template-compiler: remote js workflow
 * service-templates-core, prov-template-archetype: support for sql testing sql queries and plots for developers
+* service-templates-core: return of csv format (for csv input)
 
 `2.2.3`
 * prov-template: update to instantiation